doc: move @Fishrock123 to TSC Emeriti
It was a good run. Almost 5 years.

I haven't really been involved in the last 3+? months though,
so it's time I call it and 'retire'.

I think it is unlikely that I'll be on the TSC again,
as node is unfortunately becoming increasingly
disinteresting (& frustrating) to me.

(So long and thanks for all the fish!)

PR-URL: https://github.com/nodejs/node/pull/31725
Reviewed-By: Matteo Collina <matteo.collina@gmail.com>
Reviewed-By: Daniel Bevenius <daniel.bevenius@gmail.com>
Reviewed-By: James M Snell <jasnell@gmail.com>
Reviewed-By: Myles Borins <myles.borins@gmail.com>
Reviewed-By: Anna Henningsen <anna@addaleax.net>
Reviewed-By: Anto Aravinth <anto.aravinth.cse@gmail.com>
Reviewed-By: Colin Ihrig <cjihrig@gmail.com>
Reviewed-By: Tobias Nießen <tniessen@tnie.de>
Reviewed-By: Gus Caplan <me@gus.host>
Reviewed-By: Gireesh Punathil <gpunathi@in.ibm.com>
Reviewed-By: Michael Dawson <michael_dawson@ca.ibm.com>